dom/media/tests/mochitest/test_peerConnection_bug827843.html

<!DOCTYPE HTML>
<html>
<head>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript" src="pc.js"></script>
</head>
<body>
<pre id="test">
<script type="application/javascript">
  createHTML({
    bug: "827843",
    title: "Ensure that localDescription and remoteDescription are null after close"
  });

  var steps = [
    [
      "CHECK_FOR_SANE_SDP",
      function (test) {
        // TODO: We might want to move those checks into the default chain
        ok(test.pcLocal.localDescription,
           "test.pcLocal.localDescription is not null");
        is(test.pcLocal.localDescription.type, "offer",
           "test.pcLocal.localDescription type is offer");
        ok(test.pcLocal.localDescription.sdp.length > 10,
           "test.pcLocal.localDescription body length is plausible");

        ok(test.pcLocal.remoteDescription,
           "test.pcLocal.remoteDescription is not null");
        is(test.pcLocal.remoteDescription.type, "answer",
            "test.pcLocal.remoteDescription type is answer");
        ok(test.pcLocal.remoteDescription.sdp.length > 10,
           "test.pcLocal.remoteDescription body length is plausible");

        ok(test.pcRemote.localDescription,
           "test.pcRemote.localDescription is not null");
        is(test.pcRemote.localDescription.type, "answer",
           "test.pcRemote.localDescription type is answer");
        ok(test.pcRemote.localDescription.sdp.length > 10,
           "test.pcRemote.localDescription body length is plausible");

        ok(test.pcRemote.remoteDescription,
           "test.pcRemote.remoteDescription is not null");
        is(test.pcRemote.remoteDescription.type, "offer",
           "test.pcRemote.remoteDescription type is offer");
        ok(test.pcRemote.remoteDescription.sdp.length > 10,
           "test.pcRemote.remoteDescription body length is plausible");

        test.next();
      }
    ], [
      "CHECK_SDP_ON_CLOSED_PC",
      function (test) {
        var description;
        var exception = null;

        test.pcLocal.close();

        try { description = test.pcLocal.localDescription; } catch (e) { exception = e; }
        ok(exception, "Attempt to access localDescription of pcLocal after close throws exception");
        exception = null;

        try { description = test.pcLocal.remoteDescription; } catch (e) { exception = e; }
        ok(exception, "Attempt to access remoteDescription of pcLocal after close throws exception");
        exception = null;

        test.pcRemote.close();

        try  { description = test.pcRemote.localDescription; } catch (e) { exception = e; }
        ok(exception, "Attempt to access localDescription of pcRemote after close throws exception");
        exception = null;

        try  { description = test.pcRemote.remoteDescription; } catch (e) { exception = e; }
        ok(exception, "Attempt to access remoteDescription of pcRemote after close throws exception");

        test.next();
      }
    ]
  ];

  var test;
  runTest(function () {
    test = new PeerConnectionTest();
    test.setMediaConstraints([{audio: true}], [{audio: true}]);
    test.chain.append(steps);
    test.run();
  }, true);
</script>
</pre>
</body>
</html>
